Introduce new `test_type_checker` script to test sorbet type-checker behaviour.
The script will go through files in the `test/test_data` folder looking for (`<filename>.rb`, `<filename>.out`) pairs and check that the output of running `srb tc` against the `.rb` file matches the content of the `.out` file.

An example `success.rb` is included in this commit for helping with the review.