Remove nilability for Payload and Error.
We don't need to assume Payload and Error will be nilable, and instead leave this in the hands of the user that can specify nilable types via the generic interface.
As a result of this removal, we don't need the `payload!` method or the related errors (`NilPayloadError` and `NoPayloadOnFailureError`).

The side-effect of this change is that we now can't default `payload` or `error` to `nil` in the classes initializer.
To solve this issue, a new `.blank` method is introduced in both that allows to create an empty `Success` or `Failure`.

Fixes #10