Fix test_type_checker.sh return type consistency.
`test_type_checker.sh` loops through the files in `test/test_data` and runs `srb tc` on each of them before comparing the output with the expected output using `diff`.

The return status of the script is currently the return status of the last iteration, which means if a `diff` command fails it won't be spot unless it's the last one.

This commit addressed that, by introducing a `fail` variable that keep tracks if ANY `diff` command fails, and reflects that in the overall return status.